Replace nearly all uses of `FunctionId` with `&'static NativeFunction`
All callers have trivial access to the static reference and many runtime methods do multiple lookups.  In theory there is a risk that some popular structs get a bit larger to account for the full pointer instead of the `u32`, however both  `LocalTaskType` and `CachedTaskType` stay the same size since there was preexisting slack in the struct due to alignment.